add binary overtime variable to solver
Increase the ability for the solver to quickly find a feasible solution by proposing a  binary 'íf'variable for each employee per week. This switches on ONLY when the employee maxes out on their overtime hours.  This closes off the feasible space and allows for a solution to be found on small problems